% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_delete_inverte.R
\name{AddRowColInverse}
\alias{AddRowColInverse}
\title{Obtain the inverse of a matrix with some rows/columns added}
\usage{
AddRowColInverse(B, K, C)
}
\arguments{
\item{B}{a squared matrix of size \code{p}, representing the inverse of a matrix;}

\item{K}{the matrix of size \verb{p x k} added to the former matrix;}

\item{C}{a \verb{k x k} matrix added to the former matrix;}
}
\description{
Obtain the inverse of a matrix with some rows/columns added
}
