% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CER.R
\name{CER}
\alias{CER}
\title{Classification Error Rate (CER)}
\usage{
CER(reference, estimate)
}
\arguments{
\item{reference}{the vector containing the reference classification.}

\item{estimate}{the vector containing the estimated classification.}
}
\value{
The function returns the CER. The closer is the value to 0,
the better is the estimated classification with respect to the reference one.
}
\description{
This function returns the classification error rate (CER) as a measure of the discrepancy between the reference and the estimated classifications.
}
