% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICL.R
\name{ICL}
\alias{ICL}
\title{SpaRTaCo ICL}
\usage{
ICL(x)
}
\arguments{
\item{x}{an object of class \code{spartaco}.}
}
\value{
The ICL of the model as described by \emph{Sottosanti and Risso (2021+)}.
}
\description{
This function returns the integrated completed log-likelihood (ICL, \emph{Biernacki et al., 2000}) of a SpaRTaCo model.
}
\references{
Biernacki, C., Celeux, G. and Govaert, G. (2000). Assessing a mixture model for clustering with the
integrated completed likelihood. \emph{IEEE transactions on pattern analysis and machine intelligence} \strong{22} 719–725.

Sottosanti, A. and Risso, D. (2021+) Co-clustering of Spatially Resolved Transcriptomic Data \href{https://arxiv.org/abs/2110.04872}{(preprint)}
}
