% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetropolisAllocation.R
\name{MetropolisAllocation_fast}
\alias{MetropolisAllocation_fast}
\title{Run the SpaRTaCo SE Step, speed-up version}
\usage{
MetropolisAllocation_fast(
  x,
  Cs,
  Ds,
  Uglob,
  Dglob,
  Dist,
  Mu,
  Tau,
  Xi,
  Alpha,
  Beta,
  Phi,
  maxit = 10,
  min.obs = 3,
  prob.m = c(0.7, 0.2, 0.1)
)
}
\arguments{
\item{x}{the data matrix of size \code{n} x \code{p};}

\item{maxit}{the number of iterations to be performed;}

\item{min.obs}{the minimum number of observations in a cluster;}

\item{prob.m}{a vector of probabilities. The element in position \code{m} is the probability to perform \code{m} changes.}

\item{...}{;}
}
\description{
Run the SpaRTaCo SE Step, speed-up version
}
