% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_Kmeans_ModSel.R
\name{PCA.Kmeans.KR}
\alias{PCA.Kmeans.KR}
\title{PCA K-means model selection}
\usage{
PCA.Kmeans.KR(x, K.range, R.range)
}
\arguments{
\item{x}{the data matrix;}

\item{K.range}{the range of row clusters to evaluate;}

\item{R.range}{the range of column clusters to evaluate.}
}
\value{
a vector of length 2 containing the values of \code{K} and \code{R} selected.
}
\description{
This function returns the estimated number of row and column clusters by first applying a PCA on the data,
and then determining the best K-means model based on a change point model fitted on the total within sum of squares.
}
