% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotspartacogenes.R
\name{plot.spartaco.genes}
\alias{plot.spartaco.genes}
\title{Plot the gene-specific variances given by SpaRTaCo}
\usage{
\method{plot}{spartaco.genes}(
  x,
  r = 1:ncol(x$Expectation),
  display.plots = T,
  g = 5,
  angle = 45,
  gene.label.size = 4,
  hjust = 0,
  vjust = 0
)
}
\arguments{
\item{x}{a \code{spartaco.genes} object;}

\item{g}{if not \code{NULL}, it is the number of genes in decreasing order of the expected value whose names are displayed.}

\item{use.gene.clusters}{if \code{TRUE}, it returns the plots for every row cluster.}
}
\value{
The requested plots are displayed. In addition, if assigned to an object, it will return the \code{ggplot} object.
}
\description{
This function returns the ggplots of the distribution of the gene variances estimated by SpaRTaCo.
}
