% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSpartaco.R
\name{plot.spartaco}
\alias{plot.spartaco}
\title{Plot SpaRTaCo}
\usage{
\method{plot}{spartaco}(x, type = 1, k = NULL, r = NULL, title = NULL, manual.palette = NULL, ...)
}
\arguments{
\item{x}{a \code{spartaco} object;}

\item{type}{the type of plot you want to show.
\code{1} displays the co-cluster mean levels,
\code{2} displays the co-cluster spatial signal-to-noise ratios,
\code{3} displays the map of the spots colored with respect to the estimated column clusters,
and \code{4} displays the sample mean of some or of all the gene clusters.}

\item{k}{if \code{type = 4}, it displays the sample mean of each gene cluster is shown.}

\item{r}{if \code{type = 4}, it displays the sample mean of each gene cluster is shown.}

\item{title}{the plot title. If \code{NULL}, it does not add any title.}

\item{manual.palette}{is a vector of colors used when \code{type = "spots"}.}
}
\value{
The requested plot is displayed. In addition, if assigned to an object, it will return the \code{ggplot} object.
}
\description{
This function returns the ggplots of the mean and the spatial signal-to-noise ratios from a SpaRTaCo model.
}
