% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_delete_inverte.R
\name{RemoveRowColInverse}
\alias{RemoveRowColInverse}
\title{Obtain the inverse of a matrix with some rows/columns removed}
\usage{
RemoveRowColInverse(B, j)
}
\arguments{
\item{B}{a squared matrix of size \code{p}, representing the inverse of a matrix;}

\item{j}{the indexes of the rows/columns to be removed.}
}
\description{
Obtain the inverse of a matrix with some rows/columns removed
}
