% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.spartaco}
\alias{residuals.spartaco}
\title{SpaRTaCo residuals}
\usage{
residuals.spartaco(x)
}
\arguments{
\item{x}{a \code{spartaco} object.}
}
\value{
An object of class \code{spartaco} with the same values of the input object, expect \verb{$x}, which now contains the matrix of the residuals.
}
\description{
This function the residuals of a SpaRTaCo model.
}
