% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartacomultirun.R
\name{spartaco_multirun}
\alias{spartaco_multirun}
\title{Multiple runs of SpaRTaCo}
\usage{
spartaco_multirun(
  data,
  coordinates = NULL,
  assay = NULL,
  K = NULL,
  R = NULL,
  nstart = 5,
  Delta.constr = 10,
  max.iter = 1000,
  metropolis.iterations = 150,
  estimate.iterations = 100,
  prob.m = c(0.7, 0.2, 0.1),
  conv.criterion = list(iterations = 10, epsilon = 1e-04),
  mc.cores = 1
)
}
\arguments{
\item{data}{either a \code{SpatialExperiment} object or a matrix containing the experiment;}

\item{coordinates}{if \code{is.matrix(data)}, it takes the matrix of spatial coordinates of dimension \code{ncol(data)} x 2.}

\item{assay}{if \code{class(data) == "SpatialExperiment"}, it takes either the name or the index of the assay to be used;}

\item{K}{the number of row clusters (only when \code{input.values == NULL});}

\item{R}{the number of column clusters (only when \code{input.values == NULL});}

\item{nstart}{the number of parallel runs of the estimation algorithm;}

\item{Delta.constr}{the constraint on the Delta matrix (default is 10; see \strong{Details}).}

\item{max.iter}{the maximum number of iterations the estimation algorithm is run.}

\item{metropolis.iterations}{the number of iterations within each SE Step.}

\item{estimate.iterations}{the maximum number of iterations within each M Step.}

\item{prob.m}{the vector of probabilities assigned to the vector \code{1:length(prob.m)}, determining the number of columns the algorithm attempts to reallocate in a single SE Step.}

\item{conv.criterion}{a list containing the parameters that define a converge criterion (see \strong{Details}).}

\item{mc.cores}{the number of cores to be used.}
}
\value{
An object of class \code{spartaco} with the parameter estimates, the clustering labels, the log-likelihood value at each iteration and the data, the ICL, the data matrix and the coordinates matrix, and the clustering uncertainty.
}
\description{
This function returns the estimated model parameters and the co-clustering labels obtained after running SpaRTaCo multiple times in parallel.
}
\details{
For details about the parameters, check \code{\link[=spartaco]{spartaco()}}
}
\references{
Sottosanti, A. and Risso, D. (2021+) Co-clustering of Spatially Resolved Transcriptomic Data \href{https://arxiv.org/abs/2110.04872}{(preprint)}
}
