% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneVariances.R
\name{GeneVariances}
\alias{GeneVariances}
\title{Compute the gene-specific variances}
\usage{
GeneVariances(
  results,
  gene.names = NULL,
  HPD.interval = list(level = 0.95, R = 10^4)
)
}
\arguments{
\item{results}{The output of a \code{spartaco} function;}

\item{gene.names}{A vector of length \code{nrow(results$x)}, containing the names of the genes. If \code{NULL}, the function takes \code{gene.names = row.names(results$x)}.}

\item{HPD.interval}{A list containing the options for computing the random parameters high posterior density intervals of level \code{HPD.interval$level} (default \code{0.95}).
The intervals are computed numerically via MCMC; the number of draws used to compute the intervals is given by \code{HPD.interval$level} (default is \code{10^4}).
If \code{NULL}, the intervals are not computed.}
}
\value{
A list of five elements:
\itemize{
\item \code{Expectation}: a data frame of dimension \code{nrow(results$x)} x \code{R} containing the expected values of the random effects within the \code{R} spot clusters;
\item \code{Variance}: a data frame of dimension \code{nrow(results$x)} x \code{R} containing the variance of the random effects within the \code{R} spot clusters;
\item \code{HPD.left}: a data frame of dimension \code{nrow(results$x)} x \code{R} containing the left HPD interval of the random effects within the \code{R} spot clusters (only if \code{!is.null(HPD.interval)});
\item \code{HPD.right} a data frame of dimension \code{nrow(results$x)} x \code{R} containing the right HPD interval of the random effects within the \code{R} spot clusters (only if \code{!is.null(HPD.interval)});
\item \code{Cs}: the row clustering labels (they are the same contained into \code{results$Cs}).
}

More details on how to access the random parameter estimates are given in the Examples.
}
\description{
This function returns the summary statistics for the random variances estimated through SpaRTaCo within the \code{K*R} blocks.
}
\examples{
library(spartaco)
# Not run:
# x <- matrix(runif(n*p), n, p)
# coordinates <- matrix(runif(2*p), p, 2)
# output <- spartaco(x = x, coordinates = coordinates, K = K, R = R)

sigma2 <- GeneVariances(output)

# To access the expectations of genes within the kr-th block:
sigma2$Expectation[sigma2$Cs == k,r]

# To access the variances of the genes within the kr-th block:
sigma2$Variance[sigma2$Cs == k,r]

# To access the HPD interval limits of the genes within the kr-th block:
sigma2$HPD.left[sigma2$Cs == k,r]
sigma2$HPD.right[sigma2$Cs == k,r]
}
