% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICL.R
\name{ICL}
\alias{ICL}
\title{SpaRTaCo ICL}
\usage{
ICL(x)
}
\arguments{
\item{x}{a \code{spartaco} object.}
}
\value{
The function returns the ICL of the model as described by Sottosanti and Risso (2021+).
}
\description{
This function returns the ICL of a SpaRTaCo model.
}
