% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logL_sparseBC.R
\name{logL.sparseBC}
\alias{logL.sparseBC}
\title{Log-likelihood and the ICL of the Sparse Biclustering model}
\usage{
logL.sparseBC(
  x,
  Cs = NULL,
  Ds = NULL,
  mu = NULL,
  lambda = 0,
  results = NULL,
  ICL = F
)
}
\arguments{
\item{x}{the data matrix;}

\item{Cs}{the row clustering labels;}

\item{Ds}{the column clustering labels;}

\item{mu}{the \code{K}x\code{R} matrix containing the co-cluster centroids;}

\item{lambda}{the penalization parameter;}

\item{results}{it recieves an object of class \code{sparseBC} as input;}

\item{ICL}{logical; if true, it returns the integrated complete log-likelihood of the model.}
}
\value{
If \code{ICL = F}, it returns the log-likelihood of the model, otherwise it returns the ICL.
}
\description{
This function computes the log-likelihood and the ICL of the Sparse Biclustering model of Tan & Witten (2014).
}
