% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logL_sparseBC.R
\name{logL.sparseBC}
\alias{logL.sparseBC}
\title{Log-likelihood and ICL of the Sparse Biclustering model}
\usage{
logL.sparseBC(x, results = NULL, Cs = NULL, Ds = NULL, mu = NULL, lambda = 0)
}
\arguments{
\item{x}{the data matrix;}

\item{results}{an object of class \code{sparseBC};}

\item{Cs}{the row clustering labels (only if \code{results = NULL});}

\item{Ds}{the column clustering labels (only if \code{results = NULL});}

\item{mu}{the \code{K}x\code{R} matrix containing the co-cluster centroids (only if \code{results = NULL});}

\item{lambda}{the penalization parameter (only if \code{results = NULL});}
}
\value{
A list containing the log-likelihood, the ICL and the ML estimate of the scale parameter.
}
\description{
This function computes the log-likelihood and the ICL of the Sparse Biclustering model of Tan & Witten (2014).
}
