% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_Kmeans_ModSel.R
\name{PCA.Kmeans.KR}
\alias{PCA.Kmeans.KR}
\title{PCA K-means model selection}
\usage{
PCA.Kmeans.KR(x, K.max, R.max, continuity = F)
}
\arguments{
\item{x}{the data matrix;}

\item{K.max}{the maximum number of row clusters to evaluate;}

\item{R.max}{the maximum number of column clusters to evaluate.}

\item{continuity}{if \code{TRUE}, the change point linear model is fitted under the constraint that the function is continuous in the change point.}
}
\value{
A list containing the values of \code{K} and \code{R} selected and the clustering labels with the
selected dimensions.
}
\description{
This function returns the estimated number of row and column clusters by first applying a PCA on the data,
and then determining the best K-means model based on a change point model fitted on the total within sum of squares.
}
