% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSpartaco.R
\name{plot.spartaco}
\alias{plot.spartaco}
\title{Plot SpaRTaCo}
\usage{
\method{plot}{spartaco}(x, type = c("mean", "stn-ratio", "spots"), title = NULL)
}
\arguments{
\item{x}{a \code{spartaco} object;}

\item{type}{the type of plot you want to show. To display the co-cluster mean levels, set it to \code{"mean"}.
To display the co-cluster spatial signal-to-noise ratios, set it to \code{"stn-ratio"}.
To display the map of the spots coloured with respect to the estimated column clusters, set it to \code{"spots"}.}

\item{title}{the plot title. If \code{NULL}, it does not add any title.}
}
\value{
THe requested plot is displayed. In addition, if assigned to an object, it will return the \code{ggplot} object.
}
\description{
This function returns the ggplots of the mean and the spatial signal-to-noise ratios from a SpaRTaCo model.
}
