% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSpartaco.R
\name{plot.spartaco}
\alias{plot.spartaco}
\title{Plot SpaRTaCo}
\usage{
\method{plot}{spartaco}(
  x,
  type = 1,
  gene.name = readline("gene name: "),
  k = NULL,
  r = 1:ncol(x$mu),
  manual.palette = NULL,
  display.all.spots = T,
  ...
)
}
\arguments{
\item{x}{a \code{spartaco} object;}

\item{type}{the type of plot you want to show.
\itemize{
\item \code{1} displays the co-cluster mean levels;
\item \code{2} displays the co-cluster spatial signal-to-noise ratios;
\item \code{3} displays the map of the spots colored with respect to the estimated column clusters;
\item \code{4} if \code{!is.null(k)}, it displays the average gene expression in the clusters \code{k} and \code{r}, otherwise it displays the expression of the gene given in \code{gene.name}.
}}

\item{gene.name}{(used only when \verb{type == 4)}; it receives the name of the gene to display. If \code{!is.null(k)}, it displays the average expression of the gene clusters given by \code{k}.}

\item{k}{(used only when \verb{type == 4)} the gene clusters to plot.}

\item{r}{(used only when \verb{type == 4)} the spot clusters to plot.}

\item{manual.palette}{a vector of colors used when \code{type == 3}.}

\item{display.all.spots}{if \code{TRUE} (default) and \code{type == 4}, it displays the entire grid of spots.}
}
\value{
The requested plot is displayed. In addition, if assigned to an object, it will return the \code{ggplot} object.
}
\description{
This function returns the ggplots of the mean and the spatial signal-to-noise ratios from a SpaRTaCo model.
}
