% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartaco.R
\name{spartaco}
\alias{spartaco}
\title{SpaRTaCo co-clustering}
\usage{
spartaco(
  data,
  coordinates = NULL,
  assay = NULL,
  K = NULL,
  R = NULL,
  Delta.constr = 10,
  max.iter = 1000,
  metropolis.iterations = 150,
  estimate.iterations = 100,
  prob.m = c(0.7, 0.2, 0.1),
  input.values = NULL,
  conv.criterion = list(iterations = 10, epsilon = 1e-04),
  verbose = T,
  verbose.parallel.label = NULL,
  save.options = NULL,
  seed = NULL
)
}
\arguments{
\item{data}{either a \code{SpatialExperiment} object or a matrix containing the experiment;}

\item{coordinates}{if \code{is.matrix(data)}, it takes the matrix of spatial coordinates of dimension \code{ncol(data)} x 2.}

\item{assay}{if \code{class(data) == "SpatialExperiment"}, it takes either the name or the index of the assay to be used;}

\item{K}{the number of row clusters (only when \code{input.values == NULL}).}

\item{R}{the number of column clusters (only when \code{input.values == NULL}).}

\item{Delta.constr}{the constraint on the Delta matrix (default is 10; see \strong{Details}).}

\item{max.iter}{the maximum number of iterations the estimation algorithm is run.}

\item{metropolis.iterations}{the number of iterations within each SE Step.}

\item{estimate.iterations}{the maximum number of iterations within each M Step.}

\item{prob.m}{the vector of probabilities assigned to the vector \code{1:length(prob.m)}, determining the number of columns the algorithm attempts to reallocate in a single SE Step.}

\item{input.values}{the starting points of the estimation process (see \strong{Details}). If passed, the values of \code{K} and \code{R} are taken from it. The output of a previous model estimation can be passed here.}

\item{conv.criterion}{a list containing the parameters that define a converge criterion (see \strong{Details}).}

\item{verbose}{different ways to display the on-going estimation process (see \strong{Details}).}

\item{verbose.parallel.label}{an additional label that is used to display the on-going estimation process when \code{verbose == "parallel"}.}

\item{save.options}{a list for specifying the saving parameters (see \strong{Details}).}

\item{seed}{set the interval seed of the function.}
}
\value{
An object of class \code{spartaco} with the parameter estimates, the clustering labels, the log-likelihood value at each iteration and the data, the ICL, the data matrix and the coordinates matrix.
}
\description{
This function returns the estimated model parameters and the co-clustering labels.
}
\details{
\code{Delta.constr} gives the quantity \deqn{c = \tau_{kr} + \xi_{kr},} where \eqn{\tau_{kr}} and \eqn{\xi_{kr}} are the spatial variance and the nugget effect of block \eqn{(k,r)}.

The algorithm can be initiated from a given set of starting values. To do so, \code{input.values} receives a list of the form
\code{list(mu, tau, xi, alpha, beta, phi, Cs, Ds)}, where:
\itemize{
\item \code{mu}, \code{tau}, \code{xi}, \code{alpha} and \code{beta} are \code{K} x \code{R} matrices;
\item \code{phi} is a vector of length \code{R};
\item \code{Cs} is a vector of length \code{nrow(data)} containing the row clustering labels;
\item \code{Ds} is a vector of length \code{ncol(data)} containing the column clustering labels.
}

If the algorithm is initiated from some starting values,  \code{K} and \code{R} are set automatically according to the input values.
If an object of class \code{spartaco} is passed to \code{input.values}, the estimation starts from the final estimate of the previous run (see \strong{Examples}).

If \code{conv.criterion == NULL}, the algorithm is stopped after \code{max.iter} itereations, otherwise it is stopped when the increment of the log-likelihood is smaller than a certain threshold \code{conv.criterion$epsilon} for \code{conv.criterion$iterations} times in a row.

The function allows also to save the results even if the estimation is not completed. To do so, \code{save.options} receives a list of two parameters:
\code{after} gives the number of iterations after which the results are saved, \code{file.name} contains the path where the results are saved.

If \code{verbose == "full"}, the on-going estimation procedure is displayed. If \code{verbose == "progress"}, a dynamic progress bar will display the percentage of iterations completed.
If \code{verbose == F}, then nothing is displayed in console.
}
\examples{
library(spartaco)

# When x is a matrix:
x <- matrix(runif(n*p), n, p)
coordinates <- matrix(runif(2*p), p, 2)
output <- spartaco(data = x, coordinates = coordinates, K = K, R = R)

# When x is an object of class "SpatialExperiment":
library(spatialLIBD)
ehub <- ExperimentHub::ExperimentHub()
if (!exists("spe")) spe <- fetch_data(type = "spe", eh = ehub)
output <- spartaco(data = spe, assay = 2, K = K, R = R)

# To start the algorithm from the output of a previous run
output2 <- spartaco(data = x, coordinates = coordinates, input.val = output)
}
\references{
Sottosanti, A. and Risso, D. (2021+) Co-clustering of Spatially Resolved Transcriptomic Data \href{https://arxiv.org/abs/2110.04872}{(preprint)}
}
