% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartaco.R
\name{spartaco}
\alias{spartaco}
\title{Spartaco co-clustering}
\usage{
spartaco(spe, K = 2, R = 4, method = c("default", "marginal"))
}
\description{
This function will return the estimated parameters and the co-cluster labels.
}
\examples{
library(SpatialExperiment)
example(SpatialExperiment)
spartaco(se)
}
