% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spartaco.R
\name{spartaco}
\alias{spartaco}
\title{Spartaco co-clustering}
\usage{
spartaco(x, coordinates, K = 2, R = 4)
}
\description{
This function will return the estimated parameters and the co-clustering labels.
}
\values{
- `Cs`: row clustering labels
- `Ds`: column clustering labels
}
\examples{
library(SpatialExperiment)
example(SpatialExperiment)
spartaco(se)
}
